pyproject.toml
[project]
name = "spotify-manage"
version = "0.1.0"
description = ""
authors = [{ name = "Andre", email = "pretorius.dre+github@gmail.com" }]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "spotipy>=2.24.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "openpyxl>=3.1.5",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []
